requirements.txt
google-cloud-translate==2.0.1
django==1.9
Pyside
